pyproject.toml
[project]
name = "sw-nebula-service"
version = "0.1.0"
description = "This is my Python project."
authors = [{ name = "Arda Can Aras", email = "ardaaras99@gmail.com" }]
requires-python = ">=3.10,<3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
readme = "README.md"

dependencies = [
    "nebula3-python>=3.8.3",
    "pandas>=2.3.0",
    "pydantic>=2.11.7",
    "rich>=14.0.0",
    "sw-onto-generation",
]

[dependency-groups]
dev = [

    "ipykernel>=6.29.5",
    "ipython>=8.34.0",
    "ipywidgets>=8.1.5",
    "jupyter>=1.1.1",

    "pytest-cov>=6.0.0",
    "pytest>=8.3.5",
    "commitizen>=4.4.1",
    "pep8-naming>=0.14.1",
    "pre-commit>=4.2.0",
    "pyupgrade>=3.19.1",
    "ruff>=0.11.0",
    "tox-uv>=1.25.0",
]

[tool.ruff]
line-length = 200

[tool.ruff.format]
# Like Black
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
extend-select = [
    "A",  # flake8-builtins
    "B",  # flake8-bugbear
    "S",  # flake8-bandit
    "I",  # isort
    "UP", # pyupgrade
    "N",  # pep8-naming
]

[tool.ruff.lint.extend-per-file-ignores]
"tests/*" = ["S101"]

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version = "0.1.0"
version_provider = "uv"
version_files = [
    "pyproject.toml:version",
    "sw_nebula_service/__init__.py:__version__",
]
major_version_zero = true
update_changelog_on_bump = true

[tool.uv.sources]
sw-onto-generation = { git = "https://github.com/ardaaras99/sw-onto-generation.git" }
